Use SVG image for fleurons
Basically, it looks like shit in any browser other than Firefox :-(

In Chrome and Opera the right one is smaller and it looks weird. In IE
and Edge it's even worse.

I really like it like this, so hack it with SVG. This is just the left
fleuron in Inkscape mirrored and saved. I manually removed crap from the
XML file to make it as small as possible. The base64'd SVG is 3.8k,
which is still okay I guess.