Animate just the colour in links
This would cause screen flickering on Safari in iPhone SE (maybe also
other devices, didn't test).

What was going on is that this would also animate making the font
smaller in a media query on the initial load, triggering some
back-and-forth between the regular font size, and the smaller one in the
media query.

Thanks to lgregg at HN for reporting
https://news.ycombinator.com/item?id=19656074